refactor(ui): move flattenObjectPreservingWeaveTypes into Browse3
