+++ weave-js/src/components/PagePanelComponents/Home/Browse3/flattenObject.ts
