+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/LeaderboardPage/leaderboardConfigEditorHooks.ts
@@ -1,7 +1,7 @@
 import {parseRefMaybe, refUri} from '@wandb/weave/react';
 import {useEffect, useState} from 'react';
 
-import {flattenObjectPreservingWeaveTypes} from '../../../Browse2/browse2Util';
+import {flattenObjectPreservingWeaveTypes} from '../../flattenObject';
 import {ALL_VALUE} from '../../views/Leaderboard/types/leaderboardConfigType';
 import {EVALUATE_OP_NAME_POST_PYDANTIC} from '../common/heuristics';
 import {useGetTraceServerClientContext} from '../wfReactInterface/traceServerClientContext';
