+++ weave-js/src/components/PagePanelComponents/Home/Browse3/views/Leaderboard/query/leaderboardQuery.ts
@@ -1,7 +1,7 @@
 import {isWeaveObjectRef, parseRefMaybe} from '@wandb/weave/react';
 import _ from 'lodash';
 
-import {flattenObjectPreservingWeaveTypes} from '../../../../Browse2/browse2Util';
+import {flattenObjectPreservingWeaveTypes} from '../../../flattenObject';
 import {EVALUATE_OP_NAME_POST_PYDANTIC} from '../../../pages/common/heuristics';
 import {TraceServerClient} from '../../../pages/wfReactInterface/traceServerClient';
 import {TraceObjSchema} from '../../../pages/wfReactInterface/traceServerClientTypes';
